+++ clients/admin-ui/src/features/datastore-connections/datastore-connection.slice.ts
@@ -167,4 +167,5 @@ export const datastoreConnectionApi = createApi({
 export const {
   useGetAllDatastoreConnectionsQuery,
   useLazyGetDatastoreConnectionStatusQuery,
+  useDeleteDatastoreConnectionMutation,
 } = datastoreConnectionApi;